Improve data handling with async
